+++ src/API/useDownloadLinks.js
@@ -1,0 +1,18 @@
+import {
+  useQuery,
+} from 'react-query';
+
+
+import { useOkapiKy } from '@folio/stripes/core';
+
+export const useDownloadLinks = (id) => {
+  const ky = useOkapiKy();
+
+  const { data } = useQuery('error-download', { queryFn: () => {
+    return ky.get(`data-export-spring/jobs/${id}`).json();
+  } });
+
+  return {
+    data,
+  };
+};