+++ src/components/BulkEditList/BulkEditListFilters/ListSelect/ListSelect.js
@@ -1,11 +1,15 @@
+import { useMemo } from 'react';
+import { useLocation } from 'react-router-dom';
+import PropTypes from 'prop-types';
 import { FormattedMessage, useIntl } from 'react-intl';
 
-import { Select } from '@folio/stripes-acq-components';
+import { Select } from '@folio/stripes/components';
 
 import { identifierOptions } from '../../../../constants/optionsRecordIdentifiers';
 
-export const ListSelect = () => {
+export const ListSelect = ({ disabled, hanldeRecordIdentifier }) => {
   const intl = useIntl();
+  const location = useLocation();
 
   const options = identifierOptions.map((el) => ({
     value: el.value,
@@ -13,12 +17,26 @@ export const ListSelect = () => {
     disabled: el.disabled,
   }));
 
+  const defaultIdentifier = useMemo(() => {
+    const identifier = new URLSearchParams(location.search).get('identifier');
+
+    return identifier || options[0].value;
+  // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [location.search]);
+
   return (
     <Select
       dataOptions={options}
       arial-label={intl.formatMessage({ id: 'ui-bulk-edit.list.filters.recordIdentifier' })}
       label={<FormattedMessage id="ui-bulk-edit.list.filters.recordIdentifier" />}
-      defaultValue={options[0].value}
+      defaultValue={defaultIdentifier}
+      onChange={hanldeRecordIdentifier}
+      disabled={disabled}
     />
   );
+};
+
+ListSelect.propTypes = {
+  hanldeRecordIdentifier: PropTypes.func.isRequired,
+  disabled: PropTypes.bool,
 };