+++ src/components/ListFileUploader/index.js
@@ -1,0 +1,1 @@
+export { default as ListFileUploader } from './ListFileUploader';