+++ CHANGELOG.md
@@ -11,4 +11,9 @@
 * [UIBULKED-8](https://issues.folio.org/browse/UIBULKED-8) Add Saved queries accordion
 * [UIBULKED-23](https://issues.folio.org/browse/UIBULKED-23) Add "Drop to continue" label to the Drag and drop area
 * [UIBULKED-10](https://issues.folio.org/browse/UIBULKED-10) Actions menu
+* [UIBULKED-24](https://issues.folio.org/browse/UIBULKED-24) Link record identifier with the drag and drop area on the landing page
+* [UIBULKED-9](https://issues.folio.org/browse/UIBULKED-9) Uploading the file with identifiers
+* [UIBULKED-11](https://issues.folio.org/browse/UIBULKED-11) Bulk Edit app - permissions 
+* [UIBULKED-19](https://issues.folio.org/browse/UIBULKED-19) Download errors (CSV)
+* [UIBULKED-17](https://issues.folio.org/browse/UIBULKED-17) Start bulk edit
 * [UIBULKED-15](https://issues.folio.org/browse/UIBULKED-15) Add Query textarea